move activity from main app
