+++ packages/prop-house-webapp/src/pages/MainApp/MainApp.module.css
@@ -10,27 +10,6 @@
   font-size: 16px;
 }
 
-.activityContainer {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-  height: 70vh;
-  overflow-y: scroll;
-  padding-top: 12px;
-}
-.activityItem {
-  display: flex;
-  align-items: center;
-  font-weight: bold;
-  color: var(--brand-gray);
-  font-size: 16px;
-}
-.activityItem .address {
-  font-size: 16px;
-  font-weight: bold;
-  color: var(--brand-black);
-}
-
 .tabs li button {
   font-weight: bold;
   color: var(--brand-gray);
@@ -41,8 +20,4 @@
   border: none !important;
   border-bottom: 2px solid var(--brand-pink) !important;
   background-color: transparent !important;
-}
-
-.tabs li button.active {
-  color: red;
 }
